Move SearchBar to separated file
