+++ src/pages/LandingPage.tsx
@@ -1,14 +1,10 @@
-import { Button, Form, Select, Input } from "antd";
-import { SearchOutlined, EnvironmentOutlined } from "@ant-design/icons";
-import { useTranslation } from "react-i18next";
-import Voivodeship from "../models/Voivodeship";
+// LandingPage.tsx
 import { useEffect, useState } from "react";
+import { useTranslation } from "react-i18next";
 import store from "../store/RootStore";
 import { CaretakerSearchFilters } from "../types";
 import { useNavigate } from "react-router-dom";
-import DatePicker, { Value } from "react-multi-date-picker";
-import DatePanel from "react-multi-date-picker/plugins/date_panel";
-import weekends from "react-multi-date-picker/plugins/highlight_weekends";
+import SearchBar from "../components/SearchBar";
 
 const LandingPage = () => {
   const { t } = useTranslation();
@@ -25,54 +21,6 @@ const LandingPage = () => {
     store.selectedMenuOption = "home";
   }, []);
 
-  const renderSelectOptions = (options: Record<string, string>) => {
-    return Object.entries(options).map(([value, label]) => (
-      <Select.Option key={value} value={value}>
-        {label}
-      </Select.Option>
-    ));
-  };
-
-  const handleCityChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFilters((prev) => ({ ...prev, cityLike: e.target.value }));
-  };
-
-  const handleVoivodeshipChange = (value: string) => {
-    setFilters((prev) => ({ ...prev, voivodeship: value }));
-  };
-
-  const handleAnimalTypeChange = (value: string) => {
-    setFilters((prev) => ({
-      ...prev,
-      animals: [{ animalType: value, offerConfigurations: [] }],
-    }));
-  };
-
-  const formatDateTime = (date: string): string => {
-    return `${date} 00:00:00.000 +0100`;
-  };
-
-  const handleAvailabilitiesChange = (availabilities: Value[][]) => {
-    setFilters((prev) => {
-      if (!prev.animals || prev.animals.length === 0) {
-        return prev;
-      }
-  
-      return {
-        ...prev,
-        animals: [{ 
-          ...prev.animals[0],
-          offerConfigurations: [{
-            availabilities: availabilities.map((dateRange) => ({
-              availableFrom: formatDateTime(dateRange[0] as string) || "",
-              availableTo: formatDateTime(dateRange[1] as string) || "",
-            })),
-          }],
-        }],
-      };
-    });
-  };
-
   const handleSearch = async () => {
     console.log("filters", filters);
     navigate("/caretaker/search", { state: { filters } });
@@ -87,72 +35,7 @@ const LandingPage = () => {
 
       <div className="search-container">
         <h3>{t("landingPage.searchTitle")}</h3>
-        <div className="search-fields">
-          <Form layout="inline" className="search-form">
-            <Form.Item
-              layout="vertical"
-              label={t("addressDetails.city")}
-              style={{ display: "flex", flexDirection: "column" }}
-            >
-              <Input
-                prefix={<EnvironmentOutlined />}
-                placeholder={t("placeholder.city")}
-                className="input-field"
-                value={filters.cityLike}
-                onChange={handleCityChange}
-              />
-            </Form.Item>
-            <Form.Item layout="vertical" label={t("addressDetails.voivodeship")}>
-              <Select
-                placeholder={t("placeholder.voivodeship")}
-                className="input-field"
-                allowClear
-                style={{ width: "180px" }}
-                value={filters.voivodeship}
-                onChange={handleVoivodeshipChange}
-              >
-                {renderSelectOptions(Voivodeship.voivodeshipMap)}
-              </Select>
-            </Form.Item>
-
-            <Form.Item layout="vertical" label={t("caretakerSearch.animalType")}>
-              <Select
-                className="search-select"
-                placeholder={t("placeholder.animalType")}
-                onChange={handleAnimalTypeChange}
-              >
-                {renderSelectOptions({
-                  DOG: t("dog"),
-                  CAT: t("cat"),
-                  BIRD: t("bird"),
-                  REPTILE: t("reptile"),
-                  HORSE: t("horse"),
-                })}
-              </Select>
-            </Form.Item>
-
-            <Form.Item layout="vertical" label={t("date")}>
-              <DatePicker
-                value={filters.animals?.[0]?.offerConfigurations?.[0]?.availabilities.map((date) => [date.availableFrom, date.availableTo])}
-                disabled={filters.animals?.length === 0}
-                onChange={handleAvailabilitiesChange}
-                placeholder={t("placeholder.date")}
-                multiple
-                range
-                format="YYYY-MM-DD"
-                style={{ width: 185 }}
-                plugins={[
-                  weekends(),
-                  <DatePanel sort="date" style={{ width: 150 }} />
-                ]}
-              />
-            </Form.Item>
-
-            <Button className="search-button" onClick={handleSearch} icon={<SearchOutlined />}>
-              {t("search")}
-            </Button>
-          </Form>
-        </div>
+        <SearchBar filters={filters} setFilters={setFilters} handleSearch={handleSearch} />
       </div>
 
       <div className="main-image">
