+++ src/scss/pages/_landingPage.scss
@@ -41,7 +41,7 @@
 }
 
 .main-content h2 {
-  font-size: 2.5rem;
+  font-size: 2vw;
   margin-bottom: 20px;
 }
 
@@ -66,12 +66,11 @@
   box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
   width: 920px;
 
-  .search-fields {
+  .search-form {
     display: grid;
     grid-template-columns: repeat(5, 1fr);
     gap: 15px;
-
-
+  
     .rmdp-container input {
       height: 31px;
       border-color: v.$grey-very-light;
@@ -80,7 +79,7 @@
         opacity: 1; /* Ensure it doesn't fade (optional) */
       }
     }
-
+  
     .search-button {
       @extend .search-button;
       align-self: flex-end;
@@ -95,10 +94,6 @@
   grid-column: span 4;
   margin-bottom: 20px;
   font-size: 1.5rem;
-}
-
-.search-form {
-  display: contents;
 }
 
 .search-select {
